requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
python-dotenv>=1.0
yfinance>=0.2
alpaca-py>=0.13
requests>=2.31
pyarrow>=15.0
lightgbm>=4.0
torch>=2.0
